refactor: Re package structure for reduce size of zip
