+++ swagger-viewer/app-src/contentscript/__test__/Helper.ts
