+++ swagger-viewer/app-src/contentscript/data/DomRepository/fixtures/GitHubPageHtml_swagger_2_0_json.ts
