+++ swagger-viewer/app-src/contentscript/data/DomRepository/fixtures/GitHubPageHtml_swagger_3_0_yaml.ts
