+++ swagger-viewer/app-src/contentscript/data/QuerySelector/Document.ts
