+++ swagger-viewer/app-src/contentscript/data/QuerySelector/index.test.data.ts
