+++ swagger-viewer/app-src/contentscript/data/QuerySelector/index.ts
