+++ swagger-viewer/app-src/contentscript/presentation/App/fixtures/SwaggerJson_2_0.ts
