+++ swagger-viewer/app-src/contentscript/presentation/App/fixtures/SwaggerYaml_2_0.ts
