+++ swagger-viewer/app-src/contentscript/presentation/App/fixtures/SwaggerYaml_3_0.ts
