+++ swagger-viewer/app-src/contentscript/presentation/App/index.test.tsx
