+++ swagger-viewer/app-src/contentscript/util/YmlUtils.ts
