+++ swagger-viewer/app-src/shared/__test__/fixtures/swagger_2_0.json
