+++ swagger-viewer/app-src/shared/__test__/fixtures/swagger_3_0.yaml
