+++ swagger-viewer/app-src/shared/constants/App.ts
