+++ swagger-viewer/app-src/shared/constants/SendMessageTypes.ts
