+++ swagger-viewer/app-src/shared/types/SendMessage.ts
