+++ swagger-viewer/app-src/shared/types/Swagger.ts
