+++ swagger-viewer/app-src/shared/types/declare.ts
